Changed GameCard to GamesList
