+++ boardgame-ui/src/components/event/EventGamesTab.vue
@@ -8,33 +8,20 @@
             {{$t('button.add-game')}}
           </router-link>
         </p>
+        
         <p class="has-text-centered has-text-grey" v-if="games.length === 0">
           {{$t('event.games.no-games')}}
         </p>
-        <PanelList>
-          <GameCard
-            v-for="(game, index) in reverseSortedGames"
-            v-bind:key="index"
-            :game="game"/>
-        </PanelList>
+        
+        <GamesList v-if="games" :games="reverseSortedGames"/>
+
       </div>
     </div>
-
-    <ConfirmDeleteModal
-      :active="isConfirmDeleteModalActive"
-      :onDelete="deleteGame"
-      :onCancel="onCancelConfirmDeleteModal"
-      :content="$t('event.games.confirm-game-deletion')"/>
   </div>
 </template>
 
 <script>
-import PanelList from '@/components/layout/PanelList';
-import ConfirmDeleteModal from '@/components/layout/ConfirmDeleteModal';
-import GameCard from '@/components/games/GameCard';
-
-import Game from '@/utils/api/Game';
-
+import GamesList from '@/components/games/GamesList';
 import moment from 'moment-timezone';
 
 export default {
@@ -43,16 +30,12 @@ export default {
   },
 
   components: {
-    PanelList,
-    ConfirmDeleteModal,
-    GameCard
+    GamesList
   },
 
   data() {
     return {
       loading: true,
-      isConfirmDeleteModalActive: false,
-      gameToDelete: null,
       games: []
     };
   },
@@ -74,38 +57,6 @@ export default {
   methods: {
     sortByCreationDate: (g1, g2) => {
       return moment(g1.createdAt).diff(moment(g2.createdAt));
-    },
-
-    confirmDeleteGame: function(game) {
-      this.gameToDelete = game;
-      this.isConfirmDeleteModalActive = true;
-    },
-
-    onCancelConfirmDeleteModal: function() {
-      this.isConfirmDeleteModalActive = false;
-      this.gameToDelete = null;
-    },
-
-    async deleteGame() {
-      try {
-        await Game.delete(this.gameToDelete.id);
-        this.$toast.open({
-          message: this.$t('event.games.delete-game-success'),
-          type: 'is-success',
-          position: 'is-bottom'
-        });
-      }
-      catch(error) {
-        console.log(error);
-        this.$toast.open({
-          message: this.$t('event.games.delete-game-error'),
-          type: 'is-danger',
-          position: 'is-bottom'
-        });
-      }
-
-      this.onCancelConfirmDeleteModal();
-      this.reload();
     },
 
     async reload() {
