+++ boardgame-ui/src/components/timer/TimerList.vue
@@ -1,86 +1,86 @@
 <template>
   <div>
-      <div class="columns is-centered">
-        <div class="column">
-          <div class="container" v-if="timers.length === 0">
-            <p class="empty-list-message"><i18n path="timers.no-timer"/></p>
-          </div>
-          <router-link
-            v-for="timer in orderedTimers"
-            v-bind:key="timer.id"
-            :to="{name: 'timer', params: {timerid: timer.id}}"
-            class="box timer-box">
-              <article class="media">
-                <div class="media-left">
-                  <figure v-if="timer.board_game !== null" class="image is-64x64 is-rounded background" :style="{backgroundImage: `url('${timer.board_game.thumbnail}')`}"></figure>
-                  <figure v-else class="image is-64x64">
-                    <img class="is-rounded" :src="require('@/assets/stopwatch-icon.svg')" />
-                  </figure>
+    <div class="columns is-centered">
+      <div class="column">
+        <div class="container" v-if="timers.length === 0">
+          <p class="empty-list-message"><i18n path="timers.no-timer"/></p>
+        </div>
+        <router-link
+          v-for="timer in orderedTimers"
+          v-bind:key="timer.id"
+          :to="{name: 'timer', params: {timerid: timer.id}}"
+          class="box timer-box">
+            <article class="media">
+              <div class="media-left">
+                <figure v-if="timer.board_game !== null" class="image is-64x64 is-rounded background" :style="{backgroundImage: `url('${timer.board_game.thumbnail}')`}"></figure>
+                <figure v-else class="image is-64x64">
+                  <img class="is-rounded" :src="require('@/assets/stopwatch-icon.svg')" />
+                </figure>
+              </div>
+
+              <div class="media-content">
+
+                <nav class="level is-mobile-small-device header-level">
+                  <div class="level-left">
+                    <p class="is-size-7"><i18n v-bind:path="timerTypeI18nPathMap[timer.timer_type]"/> | {{timer.id}}</p>
+                  </div>
+                  
+                  <div class="level-left">
+                    <!-- To be re-worked in 3.1, edit is not possible right now
+                    <router-link
+                      :to="{name: 'edit-timer', params: {id: timer.id}}"
+                      class="level-item">
+                      <span class="icon is-small has-text-info">
+                        <i class="far fa-edit"></i>
+                      </span>
+                    </router-link>
+                    -->
+                    <a class="level-right" @click.prevent.stop="triggerConfirmDeleteModal(timer.id)">
+                      <span class="icon is-small has-text-danger">
+                        <i class="far fa-trash-alt"></i>
+                      </span>
+                    </a>
+                  </div>
+                </nav>
+
+                <div class="content players-list">
+                  <b-taglist>
+                    <b-tag
+                      rounded
+                      type="is-info"
+                      size="is-small"
+                      v-for="player in timer.player_timers"
+                      v-bind:key="player.id">{{getPlayerName(player)}}</b-tag>
+                  </b-taglist>
                 </div>
 
-                <div class="media-content">
-
-                  <nav class="level is-mobile-small-device header-level">
-                    <div class="level-left">
-                      <p class="is-size-7"><i18n v-bind:path="timerTypeI18nPathMap[timer.timer_type]"/> | {{timer.id}}</p>
-                    </div>
-                    
-                    <div class="level-left">
-                      <!-- To be re-worked in 3.1, edit is not possible right now
-                      <router-link
-                        :to="{name: 'edit-timer', params: {id: timer.id}}"
-                        class="level-item">
-                        <span class="icon is-small has-text-info">
-                          <i class="far fa-edit"></i>
-                        </span>
-                      </router-link>
-                      -->
-                      <a class="level-right" @click.prevent.stop="triggerConfirmDeleteModal(timer.id)">
-                        <span class="icon is-small has-text-danger">
-                          <i class="far fa-trash-alt"></i>
-                        </span>
-                      </a>
-                    </div>
-                  </nav>
-
-                  <div class="content players-list">
-                    <b-taglist>
-                      <b-tag
-                        rounded
-                        type="is-info"
-                        size="is-small"
-                        v-for="player in timer.player_timers"
-                        v-bind:key="player.id">{{getPlayerName(player)}}</b-tag>
-                    </b-taglist>
+                <nav class="level is-mobile-small-device footer-level">
+                  <div class="level-left">
+                    <p class="is-size-7">{{capitalizeFirstLetter(elapsedFromNow(mostRecentPlayer(timer.player_timers).updatedAt))}}</p>
                   </div>
 
-                  <nav class="level is-mobile-small-device footer-level">
-                    <div class="level-left">
-                      <p class="is-size-7">{{capitalizeFirstLetter(elapsedFromNow(mostRecentPlayer(timer.player_timers).updatedAt))}}</p>
+                  <div class="level-right">
+                      <p class="is-size-7"><i18n path="timers.created-by"/> <span class="name"> {{isCurrentUserTimerCreator(timer.creator.id) ? $t('timers.creator.you') : timer.creator.name}}</span></p>
                     </div>
+                </nav>
+              </div>
 
-                    <div class="level-right">
-                        <p class="is-size-7"><i18n path="timers.created-by"/> <span class="name"> {{isCurrentUserTimerCreator(timer.creator.id) ? $t('timers.creator.you') : timer.creator.name}}</span></p>
-                      </div>
-                  </nav>
-                </div>
-
-                <div class="media-right is-vertical-center is-media-right arrow-right">
-                  <span>
-                    <i class="fas fa-chevron-right"></i>
-                  </span>
-                </div>
-              </article>
-          </router-link>
-        </div>
+              <div class="media-right is-vertical-center is-media-right arrow-right">
+                <span>
+                  <i class="fas fa-chevron-right"></i>
+                </span>
+              </div>
+            </article>
+        </router-link>
       </div>
+    </div>
 
-      <confirm-delete-modal
-        :active="isConfirmDeleteModalActive"
-        :onDelete="deleteTimer"
-        :onCancel="onCancelConfirmDeleteModal"
-        :content="$t('timers.confirm-timer-deletion')"
-      />
+    <confirm-delete-modal
+      :active="isConfirmDeleteModalActive"
+      :onDelete="deleteTimer"
+      :onCancel="onCancelConfirmDeleteModal"
+      :content="$t('timers.confirm-timer-deletion')"
+    />
   </div>
 </template>
